requirements.txt
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.4
boto==2.49.0
boto3==1.16.14
botocore==1.19.14
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2
django-allauth==0.42.0
django-bootstrap-pagination==1.7.1
django-braces==1.14.0
django-ckeditor==6.0.0
django-cors-headers==2.2.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-pagination-bootstrap==1.3
django-rest-auth==0.9.5
django-social-share==1.4.0
django-storages==1.10.1
django-tinymce4-lite==1.8.0
django-widget-tweaks==1.4.8
djangorestframework==3.8.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.6.0
importlib-resources==1.5.0
isort==5.6.4
jmespath==0.10.0
jsmin==2.2.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.0.1
pipenv==2018.11.26
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
toml==0.10.1
twilio==6.45.4
urllib3==1.25.10
virtualenv==20.1.0
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
wrapt==1.12.1
zipp==1.2.0
